code polishing and scaladoc
