+++ avldb/src/main/scala/scorex/crypto/authds/avltree/batch/VersionedLDBAVLStorage.scala
@@ -1,17 +1,16 @@
 package scorex.crypto.authds.avltree.batch
 
 import com.google.common.primitives.Ints
-import scorex.core.serialization.{ErgoSerializer, ManifestSerializer}
+import scorex.core.serialization.ManifestSerializer
 import scorex.crypto.authds.avltree.batch.Constants.{DigestType, HashFnType, hashFn}
 import scorex.crypto.authds.avltree.batch.VersionedLDBAVLStorage.{topNodeHashKey, topNodeHeightKey}
 import scorex.crypto.authds.avltree.batch.serialization.{BatchAVLProverManifest, BatchAVLProverSubtree, ProxyInternalNode}
-import scorex.crypto.authds.{ADDigest, ADKey, ADValue, Balance}
+import scorex.crypto.authds.{ADDigest, ADKey}
 import scorex.util.encode.Base58
 import scorex.crypto.hash
 import scorex.crypto.hash.Digest32
 import scorex.db.{LDBKVStore, LDBVersionedStore}
 import scorex.util.ScorexLogging
-import scorex.util.serialization.{Reader, Writer}
 
 import scala.collection.mutable
 import scala.util.{Failure, Try}
@@ -35,8 +34,8 @@ class VersionedLDBAVLStorage(store: LDBVersionedStore)
   }
 
   /**
-   * Restore prover's tree from database with just pruned top node (so only one node is fetched)
-   */
+    * Restore prover's tree from database with just pruned top node (so only one node is fetched)
+    */
   def restorePrunedProver(): Try[BatchAVLProver[DigestType, HashFnType]] = {
     restorePrunedRootNode().map { recoveredTop =>
       new BatchAVLProver(StateTreeParameters.keySize, StateTreeParameters.valueSize, Some(recoveredTop))(hashFn)
@@ -61,7 +60,7 @@ class VersionedLDBAVLStorage(store: LDBVersionedStore)
                                                           additionalData: Seq[(K, V)]): Try[Unit] = {
     val digestWrapper = prover.digest
     val indexes = Seq(topNodeHashKey -> nodeLabel(prover.topNode),
-                      topNodeHeightKey -> Ints.toByteArray(prover.rootNodeHeight))
+      topNodeHeightKey -> Ints.toByteArray(prover.rootNodeHeight))
     val toInsert = serializedVisitedNodes(prover.topNode, isTop = true)
     val toRemove = prover.removedNodes().map(rn => rn.label)
     val toUpdate = indexes ++ toInsert
@@ -94,7 +93,8 @@ class VersionedLDBAVLStorage(store: LDBVersionedStore)
   /**
     * Dump current state of AVL+ tree into another (non-versioned) storage. Tree is dumped in
     * (manifest, subtrees) format
-    * @param dumpStorage - non-versioned storage to dump tree to
+    *
+    * @param dumpStorage   - non-versioned storage to dump tree to
     * @param manifestDepth - depth of manifest tree
     * @return - hash of root node of tree
     */
@@ -132,7 +132,7 @@ class VersionedLDBAVLStorage(store: LDBVersionedStore)
             manifestLoop(in.leftLabel, level + 1, manifestBuilder)
             manifestLoop(in.rightLabel, level + 1, manifestBuilder)
           case _ =>
-            // do nothing for a leaf
+          // do nothing for a leaf
         }
       }
 
@@ -156,13 +156,17 @@ class VersionedLDBAVLStorage(store: LDBVersionedStore)
 
 object VersionedLDBAVLStorage {
 
-  val topNodeHashKey: Array[Byte] = Array.fill(StateTreeParameters.labelSize)(123: Byte)
-  val topNodeHeightKey: Array[Byte] = Array.fill(StateTreeParameters.labelSize)(124: Byte)
+  private[batch] val topNodeHashKey: Array[Byte] = Array.fill(StateTreeParameters.labelSize)(123: Byte)
+  private[batch] val topNodeHeightKey: Array[Byte] = Array.fill(StateTreeParameters.labelSize)(124: Byte)
 
   // prefixes used to encode node type (internal or leaf) in database
   private[batch] val InternalNodePrefix: Byte = 0: Byte
   private[batch] val LeafPrefix: Byte = 1: Byte
 
+  /**
+    * Node serializer which is restoring nodes with only children labels stored in, without possibility to restore
+    * children automatically when requested
+    */
   val noStoreSerializer = new ProverNodeSerializer(store = null)
 
   /**
@@ -185,11 +189,19 @@ object VersionedLDBAVLStorage {
     */
   private[batch] def nodeLabel(node: ProverNodes[DigestType]): Array[Byte] = node.label
 
+  /**
+    * Reconstruct versioned AVL+ tree storage from dump
+    * @param manifest - tree started from root node cut at fixed depth
+    * @param chunks - iterator (lazy sequence) of subtrees under manifest tree
+    * @param additionalData - additional fields (metadata) to write into database
+    * @param store - db to store restored tree to
+    * @return db with reconstructed tree written to, or error
+    */
   def recreate(manifest: BatchAVLProverManifest[DigestType],
-                               chunks: Iterator[BatchAVLProverSubtree[DigestType]],
-                               additionalData: Iterator[(Array[Byte], Array[Byte])],
-                               store: LDBVersionedStore): Try[VersionedLDBAVLStorage] = {
-    //todo: the function below copy-pasted from BatchAVLProver, eliminate boilerplate
+               chunks: Iterator[BatchAVLProverSubtree[DigestType]],
+               additionalData: Iterator[(Array[Byte], Array[Byte])],
+               store: LDBVersionedStore): Try[VersionedLDBAVLStorage] = {
+    //todo: the function below copy-pasted from BatchAVLProver, eliminate boilerplate?
 
     def idCollector(node: ProverNodes[DigestType],
                     acc: Iterator[(Array[Byte], Array[Byte])]): Iterator[(Array[Byte], Array[Byte])] = {
@@ -210,7 +222,7 @@ object VersionedLDBAVLStorage {
     val indices = Iterator(topNodeHashKey -> nodeLabel(rootNode), topNodeHeightKey -> Ints.toByteArray(rootNodeHeight))
     val nodesIterator = idCollector(manifest.root, Iterator.empty) ++
       chunks.flatMap(subtree => idCollector(subtree.subtreeTop, Iterator.empty))
-    store.update(digestWrapper, toRemove = Nil, toUpdate = indices ++ nodesIterator ++ additionalData).map{_ =>
+    store.update(digestWrapper, toRemove = Nil, toUpdate = indices ++ nodesIterator ++ additionalData).map { _ =>
       new VersionedLDBAVLStorage(store)
     }
   }
@@ -226,61 +238,6 @@ object VersionedLDBAVLStorage {
     // The reason rootNodeHeight should never be more than 255 is that if height is more than 255,
     // then the AVL tree has at least  2^{255/1.4405} = 2^177 leaves, which is more than the number of atoms on planet Earth.
     ADDigest @@ (rootNodeLabel :+ rootNodeHeight.toByte)
-  }
-
-}
-
-class ProverNodeSerializer(store: LDBVersionedStore) extends ErgoSerializer[ProverNodes[DigestType]] {
-
-  import VersionedLDBAVLStorage.{InternalNodePrefix,LeafPrefix}
-
-  override def serialize(node: ProverNodes[DigestType], w: Writer): Unit = {
-    node match {
-      case n: ProxyInternalNode[DigestType] =>
-        w.put(InternalNodePrefix)
-        w.put(n.balance)
-        w.putBytes(n.key)
-        w.putBytes(n.leftLabel)
-        w.putBytes(n.rightLabel)
-      case n: InternalProverNode[DigestType] =>
-        w.put(InternalNodePrefix)
-        w.put(n.balance)
-        w.putBytes(n.key)
-        w.putBytes(n.left.label)
-        w.putBytes(n.right.label)
-      case n: ProverLeaf[DigestType] =>
-        w.put(LeafPrefix)
-        w.putBytes(n.key)
-        w.putBytes(Ints.toByteArray(n.value.length))
-        w.putBytes(n.value)
-        w.putBytes(n.nextLeafKey)
-    }
-  }
-
-  override def parse(r: Reader): ProverNodes[DigestType] = {
-    val prefix = r.getByte()
-    prefix match {
-      case InternalNodePrefix =>
-        val balance = Balance @@ r.getByte()
-        val key = ADKey @@ r.getBytes(StateTreeParameters.keySize)
-        val leftKey = ADKey @@ r.getBytes(StateTreeParameters.labelSize)
-        val rightKey = ADKey @@ r.getBytes(StateTreeParameters.labelSize)
-
-        if (store != null) {
-          new ProxyInternalProverNode(key, leftKey, rightKey, balance)(store)
-        } else {
-          new ProxyInternalNode[DigestType](key, Digest32 @@ leftKey, Digest32 @@ rightKey, balance)(hashFn)
-        }
-      case LeafPrefix =>
-        val key = ADKey @@ r.getBytes(StateTreeParameters.keySize)
-        val (value, nextLeafKey) =  {
-          val valueSize = Ints.fromByteArray(r.getBytes(4))
-          val value = ADValue @@ r.getBytes(valueSize)
-          val nextLeafKey = ADKey @@ r.getBytes(StateTreeParameters.keySize)
-          value -> nextLeafKey
-        }
-        new ProverLeaf[DigestType](key, value, nextLeafKey)(hashFn)
-    }
   }
 
 }
