+++ src/main/scala/org/ergoplatform/settings/ChainSettings.scala
@@ -2,7 +2,7 @@ package org.ergoplatform.settings
 
 import org.ergoplatform.ErgoAddressEncoder
 import org.ergoplatform.mining.AutolykosPowScheme
-import org.ergoplatform.mining.difficulty.RequiredDifficulty
+import org.ergoplatform.mining.difficulty.DifficultySerializer
 import org.ergoplatform.mining.emission.EmissionRules
 import scorex.crypto.authds.ADDigest
 import scorex.util.ModifierId
@@ -42,7 +42,7 @@ case class ChainSettings(protocolVersion: Byte,
   val initialDifficulty: BigInt = Base16.decode(initialDifficultyHex)
     .fold(_ => throw new Error(s"Failed to parse initialDifficultyHex = $initialDifficultyHex"), BigInt(_))
 
-  val initialNBits: Long = RequiredDifficulty.encodeCompactBits(initialDifficulty)
+  val initialNBits: Long = DifficultySerializer.encodeCompactBits(initialDifficulty)
 
   val initialDifficultyVersion2: BigInt = Base16.decode(voting.version2ActivationDifficultyHex)
     .fold(_ => throw new Error(s"Failed to parse initialDifficultyVersion2 = $initialDifficultyHex"), BigInt(_))
