--- src/main/scala/scorex/core/serialization/ScorexSerializer.scala
@@ -1,29 +1,0 @@
-package scorex.core.serialization
-
-import java.nio.ByteBuffer
-import scorex.util.ByteArrayBuilder
-import scorex.util.serialization._
-
-import scala.util.Try
-
-/**
-  * Basic interface for serializer with additional methods to work with bytes, not only Reader/Writer instances
-  */
-trait ErgoSerializer[T] extends Serializer[T, T, Reader, Writer] {
-
-  def toBytes(obj: T): Array[Byte] = {
-    val writer = new VLQByteBufferWriter(new ByteArrayBuilder())
-    serialize(obj, writer)
-    writer.result().toBytes
-  }
-
-  def parseBytes(bytes: Array[Byte]): T = {
-    val reader = new VLQByteBufferReader(ByteBuffer.wrap(bytes))
-    parse(reader)
-  }
-
-  def parseBytesTry(bytes: Array[Byte]): Try[T] = {
-    Try(parseBytes(bytes))
-  }
-
-}
