+++ src/test/scala/org/ergoplatform/mining/AutolykosPowSchemeSpec.scala
@@ -1,7 +1,7 @@
 package org.ergoplatform.mining
 
 import com.google.common.primitives.Ints
-import org.ergoplatform.mining.difficulty.RequiredDifficulty
+import org.ergoplatform.mining.difficulty.DifficultySerializer
 import org.ergoplatform.modifiers.history.header.{Header, HeaderSerializer}
 import org.ergoplatform.utils.ErgoPropertyTest
 import org.scalacheck.Gen
@@ -16,7 +16,7 @@ class AutolykosPowSchemeSpec extends ErgoPropertyTest with NoShrink {
     forAll(invalidHeaderGen,
             Gen.choose(100, 120),
             Gen.choose[Byte](1, 2)) { (inHeader, difficulty, ver) =>
-      val nBits = RequiredDifficulty.encodeCompactBits(difficulty)
+      val nBits = DifficultySerializer.encodeCompactBits(difficulty)
       val h = inHeader.copy(nBits = nBits, version = ver)
       val sk = randomSecret()
       val x = randomSecret()
