+++ src/test/scala/org/ergoplatform/mining/difficulty/DifficultyAdjustmentSpecification.scala
@@ -141,7 +141,7 @@ class DifficultyAdjustmentSpecification extends ErgoPropertyTest {
         val previousHeaders = control.previousHeadersRequiredForRecalculation(epoch * useLastEpochs + 1, epoch).map { i =>
           header.copy(timestamp = header.timestamp + i * interval,
             height = i,
-            nBits = RequiredDifficulty.encodeCompactBits(RequiredDifficulty.decodeCompactBits(header.nBits) + step))
+            nBits = DifficultySerializer.encodeCompactBits(DifficultySerializer.decodeCompactBits(header.nBits) + step))
         }
 
         previousHeaders.length shouldBe useLastEpochs + 1
