+++ src/test/scala/org/ergoplatform/tools/DifficultyControlSimulator.scala
@@ -1,6 +1,6 @@
 package org.ergoplatform.tools
 
-import org.ergoplatform.mining.difficulty.{DifficultyAdjustment, RequiredDifficulty}
+import org.ergoplatform.mining.difficulty.{DifficultyAdjustment, DifficultySerializer}
 import org.ergoplatform.modifiers.history.header.Header
 import org.ergoplatform.nodeView.history.ErgoHistory.Difficulty
 import org.ergoplatform.settings.ErgoSettings
@@ -70,7 +70,7 @@ object DifficultyControlSimulator extends App with ErgoGenerators {
         val timeDiff = simulateTimeDiff()
         val newHeader = lastHeader.copy(timestamp = lastHeader.timestamp + timeDiff,
           height = curHeight + 1,
-          nBits = RequiredDifficulty.encodeCompactBits(requiredDifficulty))
+          nBits = DifficultySerializer.encodeCompactBits(requiredDifficulty))
         curChain(newHeader.height) = newHeader
 
         genchain(curHeight + 1)
@@ -94,7 +94,7 @@ object DifficultyControlSimulator extends App with ErgoGenerators {
       baseHeader.copy(
         height = l(0).toInt,
         timestamp = l(1).toLong,
-        nBits = RequiredDifficulty.encodeCompactBits(BigInt(l(2)))
+        nBits = DifficultySerializer.encodeCompactBits(BigInt(l(2)))
       )
     }
     printEpochs(headers, difficultyControl)
