+++ src/test/scala/org/ergoplatform/tools/MinerBench.scala
@@ -3,7 +3,7 @@ package org.ergoplatform.tools
 import com.google.common.primitives.Bytes
 import org.bouncycastle.util.BigIntegers
 import org.ergoplatform.mining._
-import org.ergoplatform.mining.difficulty.RequiredDifficulty
+import org.ergoplatform.mining.difficulty.DifficultySerializer
 import org.ergoplatform.modifiers.history.extension.ExtensionCandidate
 import org.ergoplatform.modifiers.history.header.Header
 import org.ergoplatform.utils.ErgoTestHelpers
@@ -65,7 +65,7 @@ object MinerBench extends App with ErgoTestHelpers {
     val difficulty = 1000
     val fb = invalidErgoFullBlockGen.sample.get
     val inHeader = fb.header
-    val nBits = RequiredDifficulty.encodeCompactBits(difficulty)
+    val nBits = DifficultySerializer.encodeCompactBits(difficulty)
     val h = inHeader.copy(nBits = nBits)
 
     val candidate = new CandidateBlock(None, Header.InitialVersion, nBits: Long, h.stateRoot,
